end of video 91 (Symmetry move pt2)
This code has issues in Maya2014. Having an expression that reads directly from the transform of an object in
a container seems to cause problems with the manipulator moving that object in the scene.

Looking at the code, I can probably replace the expression with direct attribute connections. I'll give that a go
once we get to the end of the Symmetry Move feature videos.