Fixed a bug with the symmetry move feature as described in the videos.
This change provides an alternative way to copy the transformations from the selected object to its mirror
object. The technique in the videos used an expression. This uses a second locator, parent constrained to
selected object. This locator then gives you the world-space transform of the selected object. Then its
translation and rotation are just connected directly to the mirrorHelper. This approach should also be faster
than the expression. I've left both in the code so they can be switched between to see which is best.