end of video 163 (Control module duplicate)
Note, there is a bug in Maya that is making it crash when duplicating the FK control. I've added a
comment to fk.py indicating that it is the presence of the orient constraint that is the problem.
Commenting out this line allows the remainder of the duplicate module functionality to be tested.